apps/scraping/requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
soupsieve==2.8
urllib3==2.5.0
tqdm>=4.66.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0

